Share code between slice and strided slice Use same function to calculate elements in step
